Add function createStandardMethodsDecoratorFactory
