+++ lib/declarative/common/google-aip/methods/mod.ts
@@ -1,6 +1,1 @@
-export * from "./custom-method/mod.ts";
-export * from "./standard-create/mod.ts";
-export * from "./standard-delete/mod.ts";
-export * from "./standard-get/mod.ts";
-export * from "./standard-list/mod.ts";
-export * from "./standard-update/mod.ts";
+export * from "./standard-methods.ts";
