+++ lib/declarative/common/google-aip/mod.ts
@@ -1,5 +1,3 @@
 export * from "./methods/mod.ts";
 export * from "./operation.ts";
 export * from "./to-collection-identifier.ts";
-export * from "./standard-methods.ts";
-export * from "./deno-kv.ts";
