+++ lib/declarative/common/openapi/oazapfts/e2e.test.ts
@@ -1,7 +1,7 @@
 import { assertEquals } from "@std/assert";
 import { assertSnapshot } from "@std/testing/snapshot";
 import { openapi } from "#/lib/declarative/common/openapi/server.ts";
-import { standardMethodsWithDenoKv } from "#/lib/declarative/common/google-aip/deno-kv.ts";
+import { createStandardMethodsDecoratorFactory } from "#/lib/declarative/common/google-aip/methods/standard-methods.ts";
 import { createAutoSchemaDecoratorFactoryAt } from "#/lib/declarative/common/json-schema/auto-schema/auto-schema.ts";
 import { routerOf } from "#/lib/declarative/common/router/router.ts";
 import { createOazapftsClientOf, generateOazapftsClientOf } from "./client.ts";
@@ -9,10 +9,9 @@ import { createOazapftsClientOf, generateOazapftsClientOf } from "./client.ts";
 const autoSchema = await createAutoSchemaDecoratorFactoryAt(import.meta);
 
 const kv = await Deno.openKv(":memory:");
-const standardMethod = standardMethodsWithDenoKv(kv);
+const standardMethods = createStandardMethodsDecoratorFactory(kv);
 
-@standardMethod.create()
-@standardMethod.get()
+@standardMethods({ create: true, get: true })
 @autoSchema()
 class Person {
   public constructor(public name: string) {}
